requirements.txt
pydantic>=2.0
pyyaml>=6.0
google-genai>=1.0
radon>=6.0
